+++ tests/test_item.py
@@ -22,3 +22,10 @@ def test_item():
     assert_equal(i.f1, 'f1')
     assert_equal(i.f2, 'f2')
     assert_equal(i.f3, 'f3')
+
+    # change attributes
+    i.f1 = 'f1_new'
+    i.f2 = 'f2_new'
+    assert_equal(i.f1, 'f1_new')
+    assert_equal(i.f2, 'f2_new')
+    assert_equal(i.f3, 'f3')